Initial Push.
I didn't add in the csv files. I added .csv into the gitignore. Right
now it only prints the data off one file to the console. I will continue
to work on it. When running the code there needs to be a folder in the
project called data and store the csv's there.