Got the files to be read.
There were two issue. The first was the regular expression I was using.
I forgot to add '^' to the front of the regex string. This made
determening which lines to look at easier. I also changed my catch
statement so I found out that there was an out of bounds exception and
put in a change to handle that. I also am adding the csv's because our
client/professor said that its ok b/c the data is public data.